requirements.txt
streamlit
numpy
Pillow
tensorflow
pandas
matplotlib
seaborn
opencv-python
scikit-learn
keras-tuner
kagglehub 